% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_repair_path}
\alias{check_repair_path}
\title{Check URL}
\usage{
check_repair_path(path)
}
\arguments{
\item{path}{The path, for example appended to the base URL.}
}
\value{
The path string with optionally an appended trailing slash.
}
\description{
Check URL
}
