% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kobo-paginator.R
\name{KoboPaginator}
\alias{KoboPaginator}
\title{KoboPaginator}
\description{
A class that implements link-header style pagination, as is used in
the Kobotoolbox API.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{KoboPaginator$new()}}
\item \href{#method-get}{\code{KoboPaginator$get()}}
\item \href{#method-set_first_response}{\code{KoboPaginator$set_first_response()}}
\item \href{#method-get_responses}{\code{KoboPaginator$get_responses()}}
\item \href{#method-clone}{\code{KoboPaginator$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KoboPaginator$new(client)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{client}}{KoboClient. An instance of a KoboClient that can
be used for the paginated requestes.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get"></a>}}
\if{latex}{\out{\hypertarget{method-get}{}}}
\subsection{Method \code{get()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KoboPaginator$get(path, query, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_first_response"></a>}}
\if{latex}{\out{\hypertarget{method-set_first_response}{}}}
\subsection{Method \code{set_first_response()}}{
Set the initial response
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KoboPaginator$set_first_response(response)}\if{html}{\out{</div>}}
}

\subsection{Details}{
Usually, the page method would revoke a first response in the
normal way, using its \code{next} element to walk over all subsequent
pages. In some settings, the user might provide this initial response
object already.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_responses"></a>}}
\if{latex}{\out{\hypertarget{method-get_responses}{}}}
\subsection{Method \code{get_responses()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KoboPaginator$get_responses()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KoboPaginator$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
