% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_as_json_char.R
\name{list_as_json_char}
\alias{list_as_json_char}
\title{Helper function to convert R list to JSON-like string}
\usage{
list_as_json_char(list)
}
\arguments{
\item{list}{R list that should be converted to the JSON-like string}
}
\description{
Converts R lists to JSON-like strings for POST request's body.

#' @keywords internal
}
\examples{
\dontrun{
example_body <- list_as_json_char(list(
"name" = "A survey object created via API/R",
"asset_type" = "survey"))
}


}
