% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asset.R
\name{Asset}
\alias{Asset}
\title{Asset
Class to hold an asset}
\description{
Asset
Class to hold an asset

Asset
Class to hold an asset
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{uid}}{uid of the asset.}

\item{\code{name}}{name of the asset.}

\item{\code{asset_url}}{asset_url of the asset. takes the form of assets/{uid}/
Contains most of the information about an asset such as meta data, permissions, xls form content and
links to other related endpoints.}

\item{\code{data_url}}{data_url of the asset. This can be used to get data related to the asset.
Most useful for the asset type \code{survey} where this URL
gives access to the submissions to the survey.}

\item{\code{owner_username}}{username of the owner of the asset.}

\item{\code{type}}{type of the asset. Type of the asset, e.g. survey, question, block or template.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Asset$new()}}
\item \href{#method-get_submissions}{\code{Asset$get_submissions()}}
\item \href{#method-to_list}{\code{Asset$to_list()}}
\item \href{#method-clone}{\code{Asset$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Asset$new(asset_list, kobo)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{asset_list}}{list. a list with the asset data, such as obtained through a call to assets/{id}}

\item{\code{kobo}}{Kobo instance. Instance of class Kobo used internally to make requests to the API.
get_submissions}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_submissions"></a>}}
\if{latex}{\out{\hypertarget{method-get_submissions}{}}}
\subsection{Method \code{get_submissions()}}{
get submissions to a survey/form
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Asset$get_submissions()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
tibble. submissions as a tibble. if no submissions were made yet, the tibble will have no columns.
to_list
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-to_list}{}}}
\subsection{Method \code{to_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Asset$to_list()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
list representation of asset
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Asset$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
