% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kobo.R
\name{Kobo}
\alias{Kobo}
\title{Kobo Class}
\description{
Interface object for the Kobo API that can handle KoboClient instances
(sessions) for both API versions.
The Class exposes both generic and specific methods for HTTP requests /
interactions with the various endpoints.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Kobo$new()}}
\item \href{#method-get}{\code{Kobo$get()}}
\item \href{#method-post}{\code{Kobo$post()}}
\item \href{#method-get_assets}{\code{Kobo$get_assets()}}
\item \href{#method-clone_asset}{\code{Kobo$clone_asset()}}
\item \href{#method-deploy_asset}{\code{Kobo$deploy_asset()}}
\item \href{#method-create_asset}{\code{Kobo$create_asset()}}
\item \href{#method-import_xls_form}{\code{Kobo$import_xls_form()}}
\item \href{#method-clone}{\code{Kobo$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initialization method for class "Kobo".
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Kobo$new(
  base_url_v2,
  base_url_v1 = NULL,
  kobo_token = Sys.getenv("KBTBR_TOKEN")
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{base_url_v2}}{character. The base URL of the API version 2
(known as /api/v2). For example: https://kobo.correlaid.org.}

\item{\code{base_url_v1}}{character. The base URL of the API of your KoBoCAT
API (also known as /api/v1). Defaults to NULL.
For example: https://kc.correlaid.org.}

\item{\code{kobo_token}}{character. The API token. Defaults to requesting
the systen environment \code{KBTBR_TOKEN}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get"></a>}}
\if{latex}{\out{\hypertarget{method-get}{}}}
\subsection{Method \code{get()}}{
Wrapper for the GET method of internal session objects.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Kobo$get(path, query = list(format = "json"), version = "v2")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path}}{character. Path component of the endpoint.}

\item{\code{query}}{list. A named list which is parsed to the query
component. The order is not hierarchical.}

\item{\code{version}}{character. Indicates on which API version the request
should be executed (available: \code{v1}, \code{v2}). Defaults to \code{v2}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-post"></a>}}
\if{latex}{\out{\hypertarget{method-post}{}}}
\subsection{Method \code{post()}}{
Wrapper for the POST method of internal session objects.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Kobo$post(path, body, version = "v2")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path}}{character. Path component of the endpoint.}

\item{\code{body}}{R list or json-like string. A data payload to be sent
to the server.}

\item{\code{version}}{character. Indicates on which API version the request
should be executed (available: \code{v1}, \code{v2}). Defaults to \code{v2}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_assets"></a>}}
\if{latex}{\out{\hypertarget{method-get_assets}{}}}
\subsection{Method \code{get_assets()}}{
Example method to send a GET request to the \code{assets} endpoint
(due to default to \code{v2}, no further specification is needed).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Kobo$get_assets()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone_asset"></a>}}
\if{latex}{\out{\hypertarget{method-clone_asset}{}}}
\subsection{Method \code{clone_asset()}}{
High-level POST request to clone an asset. \code{assets} endpoint
(due to default to \code{v2}, no further specification is needed).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Kobo$clone_asset(clone_from, name, asset_type)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{clone_from}}{character. UID of the asset to be cloned.}

\item{\code{name}}{character. Name of the new asset.}

\item{\code{asset_type}}{character. Type of the new asset. Can be
"block", "question", "survey", "template".}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-deploy_asset"></a>}}
\if{latex}{\out{\hypertarget{method-deploy_asset}{}}}
\subsection{Method \code{deploy_asset()}}{
High-level POST request to deploy an asset.
\verb{assets/\{uid\}/deployment/} endpoint (due to
default to \code{v2}, no further specification is needed).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Kobo$deploy_asset(uid)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{uid}}{character. UID of the asset to be deployed.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-create_asset"></a>}}
\if{latex}{\out{\hypertarget{method-create_asset}{}}}
\subsection{Method \code{create_asset()}}{
High-level POST request to create an empty asset. \verb{assets/} endpoint
(due to default to \code{v2}, no further specification is needed).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Kobo$create_asset(
  name,
  description,
  sector,
  country,
  share_metadata,
  asset_type
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{character. Name of the new asset.}

\item{\code{description}}{character. Optional.}

\item{\code{sector}}{character. Optional.}

\item{\code{country}}{character. Optional.}

\item{\code{share_metadata}}{boolean. Optional.}

\item{\code{asset_type}}{character. Type of the new asset. Can be
"block", "question", "survey", "template".}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-import_xls_form"></a>}}
\if{latex}{\out{\hypertarget{method-import_xls_form}{}}}
\subsection{Method \code{import_xls_form()}}{
High-level POST request to import an XLS form. \code{imports} endpoint
(due to default to \code{v2}, no further specification is needed).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Kobo$import_xls_form(name, file)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{character. Name of the new asset.}

\item{\code{file}}{character. The path to the file containing the XLS form.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Kobo$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
